requirements.txt
alembic==1.8.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.6
django-bootstrap-icons==0.8.1
django-bootstrap-v5==1.0.11
django-cors-headers==3.13.0
django-mathfilters==1.0.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.71
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
psycopg==3.0.15
psycopg2-binary==2.8.4
pyjsparser==2.7.1
PyJWT==2.4.0
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
